// src/app/modules/cpassapi/model/interventoAltriDati.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Cpv } from './cpv';
import { Intervento } from './intervento';
import { TipoAcquisto } from './tipoAcquisto';


export interface InterventoAltriDati { 
    codiceInterno?: string;
    cpvMatRic?: Cpv;
    cpvVerdi?: Cpv;
    id?: string;
    importoIvaMatRic?: number;
    importoIvaVerdi?: number;
    importoNettoIvaMatRic?: number;
    importoNettoIvaVerdi?: number;
    importoTotMatRic?: number;
    importoTotVerdi?: number;
    intervento?: Intervento;
    ivaAnniSuccessivi?: number;
    ivaPrimoAnno?: number;
    ivaSecondoAnno?: number;
    normativaRiferimento?: string;
    note?: string;
    oggettoMatRic?: string;
    oggettoverdi?: string;
    speseSostenute?: number;
    tipoAcquistoMatRic?: TipoAcquisto;
    tipoAcquistoVerdi?: TipoAcquisto;
}